# `iterable_subprocess` based `annexworktree()`