--- a/datalad_next/iter_collections/__init__.py
+++ b/datalad_next/iter_collections/__init__.py
@@ -20,6 +20,7 @@ collections.
    annexworktree
    directory
    gitworktree
+   annexworktree
    tarfile
    zipfile
    utils