--- a/datalad_next/iter_collections/__init__.py
+++ b/datalad_next/iter_collections/__init__.py
@@ -17,6 +17,7 @@ collections.
 .. autosummary::
    :toctree: generated
 
+   annexworktree
    directory
    gitworktree
    tarfile