--- a/datalad_next/iter_collections/__init__.py
+++ b/datalad_next/iter_collections/__init__.py
@@ -20,7 +20,6 @@ collections.
    annexworktree
    directory
    gitworktree
-   annexworktree
    tarfile
    zipfile
    utils