--- a/datalad_next/iter_collections/annexworktree.py
+++ b/datalad_next/iter_collections/annexworktree.py
@@ -89,7 +89,7 @@ def iter_annexworktree(
                 # we get the annex key for any filename
                 # (or empty if not annexed)
                 ['git', '-C', str(path),
-                 'annex', 'find', '--anything', '--format=\${key}\n', '--batch'],
+                 'annex', 'find', '--anything', '--format=${key}\n', '--batch'],
                 # intersperse items with newlines to trigger a batch run
                 # this avoids string operations to append newlines to items
                 input=intersperse(