This is moving PR https://github.com/mih/datalad-next/pull/3 here, to get it tested properly. Below is the original description by @christian-monch 

It includes (sits on top of) https://github.com/datalad/datalad-next/pull/538 (merged now)

TODO

- [x] Rebase on #552
- [x] Add documentation
- [x] Make `keep_ends=True str.strip()` combo obsolete (and ease debugging why @mih cannot do it, ideally)
- [x] `fp=True` currently only does meaningful things for annexed files, but it should also act properly for any other file. This aspect of the functionality is still undertested.

----

This PR adds an implementation of `iter_annexworktree` that is based on `iterable_subprocesses` and the ideas laid out in issue https://github.com/datalad/datalad-next/issues/537.

This PR included a collection of data processors that are basically generator-wrapper.

The PR also modifies `iter_gitworktree` to use `iterable_subprocesses` instead of the datalad-core runner.

The current implementation of `iter_annexworktree` iterates over a dataset with 33k annex files in less than 5 seconds on my machine.