@mih 2023-11-22
Looking further into the timing:

Current `main`

```
>>> %timeit list(iter_gitworktree('.'))
220 ms ± 4.59 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

This PR:

```
>>> %timeit list(iter_gitworktree('.'))
221 ms ± 907 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Same. And both 5x slower than the baseline:

```
multitime -n 5 git ls-files --stage --cached --exclude-standard --others' > /dev/null
            Mean        Std.Dev.    Min         Median      Max
real        0.038       0.007       0.029       0.043       0.046       
user        0.020       0.005       0.015       0.019       0.029       
sys         0.017       0.005       0.011       0.014       0.026       
```

This explains some of the 25% overall difference, but not the 8%.