@mih 2023-11-22
After normalizing the benchmark conditions further, I have the following stats:

```
# this code
In [6]: %timeit list(iter_annexworktree('.'))
3.59 s ± 104 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# original sketch
In [7]: %timeit run_itersubproc()
4.17 s ± 317 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# baseline with iterable_subprocess
In [9]: %timeit list(iter_gitworktree('.'))
302 ms ± 8.02 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

This compared to a shell pipe (using `jq` as consumer that added the JSON decoding load):

```
❯ multitime -n 5 git ls-files | git annex find --anything --format="\${key}\n" --batch | grep -v '^[[:space:]]*$' | git annex examinekey --json --batch | jq > /dev/null
===> multitime results
1: git ls-files
            Mean        Std.Dev.    Min         Median      Max
real        3.399       0.169       3.236       3.369       3.703       
user        0.006       0.005       0.000       0.003       0.012       
sys         0.006       0.006       0.000       0.008       0.015       
```

~8% difference, and nearly within the margin of error.

This works for me!