@mih 2023-11-27
Ideally, this PR and https://github.com/datalad/datalad-next/pull/546 are now orthogonal and could be reviewed independently of each other.