@christian-monch 2023-11-27
> `align_pattern()` is introduced here (with tests), but it unused. I would prefer to keep PRs self contained. Concretely, I would prefer to not introduce `align_pattern` until it is needed, and then introduce it in that context.

Ok, I will remove it from the PR