@christian-monch 2023-11-29
> Thanks! This looks clean. My comments focus on a few aspects that may either be misunderstanding (hinting at documentation gaps), or could lead to a tuned design.
> 
> Re `iter_annexworktree()` itself: I believe it currently ignores the whole `fp=True` code branch entirely. It does the processing, but hides the outcome.

Fixed in 1e1d3e57fab42ebd1b9b20c1e6e5d91ea69473ee