@mih 2023-12-05
Current state of implementation is between 5% (Python) and 15% (CLI) slower than a rough shell equivalent:

```
❯ multitime -n 5 datalad -f disabled ls-file-collection annexworktree .
===> multitime results
1: datalad -f disabled ls-file-collection annexworktree .
            Mean        Std.Dev.    Min         Median      Max
real        3.767       0.042       3.697       3.778       3.823       
user        5.569       0.143       5.320       5.586       5.765       
sys         2.264       0.104       2.165       2.194       2.427       

❯ multitime -n 5 git ls-files | git annex find --anything --format="\${key}\n" --batch | grep -v '^[[:space:]]*$' | git annex examinekey --json --batch | jq > /dev/null
===> multitime results
1: git ls-files
            Mean        Std.Dev.    Min         Median      Max
real        3.289       0.062       3.207       3.275       3.392       
user        0.009       0.003       0.006       0.007       0.013       
sys         0.004       0.003       0.000       0.005       0.007       

In [3]: %timeit list(iter_annexworktree(Path('.')))
3.48 s ± 39.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```