@mih 2023-12-06
```
FAILED ..\iter_collections\tests\test_iterannexworktree.py::test_iter_annexworktree_basic_fp - UnicodeDecodeError: 'utf-8' codec can't decode byte 0xf6 in position 10: invalid start byte
```

Something is still broken. The test condition existed for some time, but the test never actually ran.