@mih 2023-12-06
#555 brings helpers that can be used to remove duplication of `fp=True` handling.